fix(equipments): add conflict error when deleting equipment with unclosed maintenance requests

+++ src/services/equipments.service.ts
@@ -119,6 +119,13 @@ export const deleteEquipment = (id: string): void => {
     throw new NotFoundError("Equipment is not found");
   }
 
+  const unclosedRequests = requestsService.getUnclosedRequestsByEquipmentId(id);
+  if (unclosedRequests.length > 0) {
+    throw new ConflictError(
+      "Can't remove equipment with unclosed maintenance requests",
+    );
+  }
+
   equipmentsRepository.deleteEquipment(id);
 };
